pyproject.toml
[project]
name = "filewatcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.12",
    "eventlet>=0.40.4",
    "gunicorn>=23.0.0",
    "python-socketio>=5.15.1",
    "watchdog>=6.0.0",
]
